+++ toondo/packages/presentation/lib/widgets/my_page/notification_setting/time_picker/time_picker_bottom_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:presentation/viewmodels/my_page/notification_setting/time_picker_viewmodel.dart';
-import 'package:presentation/widgets/bottom_sheet/common_bottom_sheet.dart';
+import 'package:presentation/widgets/bottom_sheet/custom_bottom_sheet.dart';
 import 'package:presentation/widgets/my_page/notification_setting/time_picker/time_picker_column.dart';
 import 'package:provider/provider.dart';
 
@@ -17,7 +17,7 @@ class TimePickerBottomSheet extends StatelessWidget {
   Widget build(BuildContext context) {
     final viewModel = context.read<TimePickerViewModel>();
 
-    return CommonBottomSheet(
+    return CustomBottomSheet(
       title: '시간 설정',
       body: Row(
         mainAxisAlignment: MainAxisAlignment.center,
