+++ toondo/packages/presentation/lib/widgets/my_page/sync_bottom_sheet.dart
@@ -1,3 +1,39 @@
+import 'package:flutter/material.dart';
+import 'package:presentation/widgets/bottom_sheet/custom_bottom_sheet.dart';
+
+class SyncBottomSheet extends StatelessWidget {
+  const SyncBottomSheet({super.key});
+
+  // TODO: 추후 ViewModel로 대체 예정
+  void _handleLoad() {
+    debugPrint('불러오기 눌림');
+  }
+
+  void _handleSave() {
+    debugPrint('저장하기 눌림');
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomBottomSheet(
+      title: '동기화 하기',
+      buttons: [
+        CommonBottomSheetButtonData(
+          label: '불러오기',
+          filled: true,
+          onPressed: _handleLoad,
+        ),
+        CommonBottomSheetButtonData(
+          label: '저장하기',
+          filled: false,
+          onPressed: _handleSave,
+        ),
+      ],
+    );
+  }
+}
+
+// -- NOTE : 아래 코드는 기존 코드로 추후 참고용으로 남겨둡니다. --
 // import 'package:flutter/material.dart';
 // import 'package:toondo/viewmodels/my_page/my_page_viewmodel.dart';
